Add a boolean toggle for gzcheck (if files are not compressed)

+++ nanite-hadoop/src/main/java/uk/bl/wap/hadoop/profiler/FormatProfiler.java
@@ -110,33 +110,41 @@ public class FormatProfiler extends Configured implements Tool {
 			System.exit( 0 );
 
 		}
+
+		String[] fpargs = new String[] { args[0], args[1]+"" };
+		int ret = 0;
+
+		final boolean gzcheck = false;
 		
-		// Run a job to check that the input warc files are ok to use
-		String[] gzargs = new String[] { args[0], args[1]+"-precheck" };
-		int ret = ToolRunner.run( new GZChecker(), gzargs);
-		
-		// Recover the output here and cache to a local file
-		FileSystem fs = FileSystem.get(new Configuration());
-		FileStatus[] i = fs.listStatus(new Path(gzargs[1]), new GlobFilter("part-*"));
-		File tempFile = File.createTempFile("nanite-", ".txt");
-		tempFile.deleteOnExit();
-		PrintWriter pw = new PrintWriter(new FileWriter(tempFile));
-		for(FileStatus f:i) {
-			BufferedReader is = new BufferedReader(new InputStreamReader(fs.open(f.getPath())));
-			String[] line = null;
-			while(is.ready()) {
-				line = is.readLine().split("\t");;
-				if(line[1].equals(GZChecker.OK)) {
-					pw.println(line[0]);
+		if(gzcheck) {
+			// Run a job to check that the input warc files are ok to use
+			final String[] gzargs = new String[] { args[0], args[1]+"-precheck" };
+			ret = ToolRunner.run( new GZChecker(), gzargs);
+
+			// Recover the output here and cache to a local file
+			FileSystem fs = FileSystem.get(new Configuration());
+			FileStatus[] i = fs.listStatus(new Path(gzargs[1]), new GlobFilter("part-*"));
+			File tempFile = File.createTempFile("nanite-", ".txt");
+			tempFile.deleteOnExit();
+			PrintWriter pw = new PrintWriter(new FileWriter(tempFile));
+			for(FileStatus f:i) {
+				BufferedReader is = new BufferedReader(new InputStreamReader(fs.open(f.getPath())));
+				String[] line = null;
+				while(is.ready()) {
+					line = is.readLine().split("\t");;
+					if(line[1].equals(GZChecker.OK)) {
+						pw.println(line[0]);
+					}
 				}
+				is.close();
 			}
-			is.close();
+			fs.close();
+			pw.close();
+
+			// Use the outputs from GZChecker to run the FormatProfiler
+			fpargs = new String[] { tempFile.getAbsolutePath(), args[1]+"" };
 		}
-		fs.close();
-		pw.close();
 		
-		// Use the outputs from GZChecker to run the FormatProfiler
-		String[] fpargs = new String[] { tempFile.getAbsolutePath(), args[1]+"" };
 		ret |= ToolRunner.run( new FormatProfiler(), fpargs );
 
 		System.exit( ret );
